+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:example/pages/custom_paint_page.dart';
+import 'package:example/pages/frame_rate_page.dart';
 import 'package:example/pages/snow_page.dart';
 import 'package:example/pages/widget_page.dart';
 import 'package:flutter/material.dart';
@@ -67,6 +68,18 @@ class MainPage extends StatelessWidget {
                 );
               },
               child: const Text('Widget Particle Demo'),
+            ),
+            const SizedBox(height: 20),
+            ElevatedButton(
+              onPressed: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => const FrameRatePage(),
+                  ),
+                );
+              },
+              child: const Text('Frame Rate Control Demo'),
             ),
           ],
         ),