+++ lib/snowfall_or_anythings.dart
@@ -1,4 +1,2 @@
-library snowfall_or_anythings;
-
 export 'src/snowfall_widget.dart';
 export 'src/particle.dart';